feat(DTFS2-6508): remove user limitation to provide feedback
