+++ gef-ui/component-tests/includes/footer.component-test.js
@@ -1,52 +1,27 @@
 const pageRenderer = require('../pageRenderer');
-const { NON_MAKER_OR_CHECKER_ROLES } = require('../../test-helpers/common-role-lists');
-const { MAKER, CHECKER } = require('../../server/constants/roles');
 
 const page = 'includes/footer.njk';
 const render = pageRenderer(page);
+const { ALL_ROLES } = require('../../test-helpers/common-role-lists');
 
 describe(page, () => {
   let wrapper;
 
-  describe(`when viewed with the role '${MAKER}'`, () => {
-    beforeEach(() => {
-      wrapper = render({
-        user: {
-          roles: [MAKER],
-        },
-      });
+  const usersWithRoles = ALL_ROLES.map((role) => ({ testDescription: role, user: { roles: [] } }));
+
+  const testCases = [...usersWithRoles, { testDescription: 'user with no roles', user: {} }, { testDescription: 'logged out user', user: { roles: [] } }];
+
+  describe.each(testCases)('viewed by a $testDescription', (role) => {
+    const user = {
+      roles: [role],
+    };
+
+    beforeAll(() => {
+      wrapper = render({ user });
     });
 
-    it('should render the feedback link', () => {
-      wrapper.expectElement('[data-cy="feedback-link"]').toExist();
-    });
-  });
-
-  describe(`when viewed with the role '${CHECKER}'`, () => {
-    beforeEach(() => {
-      wrapper = render({
-        user: {
-          roles: [CHECKER],
-        },
-      });
-    });
-
-    it('should render the feedback link', () => {
-      wrapper.expectElement('[data-cy="feedback-link"]').toExist();
-    });
-  });
-
-  describe.each(NON_MAKER_OR_CHECKER_ROLES)('when viewed with the role \'%s\'', (nonMakerOrCheckerRole) => {
-    beforeEach(() => {
-      wrapper = render({
-        user: {
-          roles: [nonMakerOrCheckerRole],
-        },
-      });
-    });
-
-    it('should NOT render the feedback link', () => {
-      wrapper.expectElement('[data-cy="feedback-link"]').notToExist();
+    it('renders a feedback link', () => {
+      wrapper.expectLink('[data-cy="feedback-link"]').toLinkTo('/feedback', 'Feedback');
     });
   });
 });
