+++ portal/component-tests/partials/footer.component-test.js
@@ -1,4 +1,4 @@
-const { MAKER, CHECKER } = require('../../server/constants/roles');
+const { ALL_ROLES } = require('../../test-helpers/common-role-lists');
 const pageRenderer = require('../pageRenderer');
 
 const page = '_partials/footer.njk';
@@ -7,9 +7,13 @@ const render = pageRenderer(page);
 describe(page, () => {
   let wrapper;
 
-  describe('viewed by a maker', () => {
+  const usersWithRoles = ALL_ROLES.map((role) => ({ testDescription: role, user: { roles: [] } }));
+
+  const testCases = [...usersWithRoles, { testDescription: 'user with no roles', user: {} }, { testDescription: 'logged out user', user: { roles: [] } }];
+
+  describe.each(testCases)('viewed by a $testDescription', (role) => {
     const user = {
-      roles: [MAKER],
+      roles: [role],
     };
 
     beforeAll(() => {
@@ -18,46 +22,6 @@ describe(page, () => {
 
     it('renders a feedback link', () => {
       wrapper.expectLink('[data-cy="feedback-link"]').toLinkTo('/feedback', 'Feedback');
-    });
-  });
-
-  describe('viewed by a checker', () => {
-    const user = {
-      roles: [CHECKER],
-    };
-
-    beforeAll(() => {
-      wrapper = render({ user });
-    });
-
-    it('renders a feedback link', () => {
-      wrapper.expectLink('[data-cy="feedback-link"]').toLinkTo('/feedback', 'Feedback');
-    });
-  });
-
-  describe('viewed by a user with no roles', () => {
-    const user = {
-      roles: [],
-    };
-
-    beforeAll(() => {
-      wrapper = render({ user });
-    });
-
-    it('does not render a feedback link', () => {
-      wrapper.expectLink('[data-cy="feedback-link"]').notToExist();
-    });
-  });
-
-  describe('viewed by a logged out user', () => {
-    const user = {};
-
-    beforeAll(() => {
-      wrapper = render({ user });
-    });
-
-    it('does not render a feedback link', () => {
-      wrapper.expectLink('[data-cy="feedback-link"]').notToExist();
     });
   });
 });
